refactor pre-processor and fix visualizer normalization issue
